Allow performing assert comparisons (e.g. ASSERT_EQ) with any traceable type
