+++ flow/include/flow/FastRef.h
@@ -24,6 +24,7 @@
 
 #include <atomic>
 #include <cstdint>
+#include "flow/Traceable.h"
 
 // The thread safety this class provides is that it's safe to call addref and
 // delref on the same object concurrently in different threads. Subclass does
@@ -178,6 +179,13 @@ public:
 
 private:
 	P* ptr;
+};
+
+template <class T>
+struct Traceable<Reference<T>> : std::conditional<Traceable<T>::value, std::true_type, std::false_type>::type {
+	static std::string toString(const Reference<T>& value) {
+		return value ? Traceable<T>::toString(*value) : "[not set]";
+	}
 };
 
 template <class P, class... Args>
